feat(package): rename app package
