--- app/src/main/java/com/example/androidcodelabapp/data/network/api/APIClient.kt
@@ -1,17 +1,0 @@
-package com.example.androidcodelabapp.data.network.api
-
-import com.example.androidcodelabapp.data.network.dto.GithubUserDto
-import com.example.androidcodelabapp.data.network.dto.GithubUsersResponseDto
-
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface APIClient {
-    @GET("search/users?q=type:User+location:Kampala+language:JAVA")
-    fun getAllDevelopers(): Call<GithubUsersResponseDto>
-
-    @GET("users/{githubUserName}")
-    fun getDeveloperProfile(@Path("githubUserName") githubUserName: String): Call<GithubUserDto>
-
-}
