--- app/src/main/java/com/example/androidcodelabapp/data/network/dto/GithubUsersResponseDto.kt
@@ -1,8 +1,0 @@
-package com.example.androidcodelabapp.data.network.dto
-
-import com.google.gson.annotations.SerializedName
-
-data class GithubUsersResponseDto(
-    @SerializedName("items")
-    val githubUsers: ArrayList<GithubUserDto> = arrayListOf()
-)
