--- app/src/main/java/com/example/androidcodelabapp/data/network/responses/NetworkResponse.kt
@@ -1,7 +1,0 @@
-package com.example.androidcodelabapp.data.network.responses
-
-data class NetworkResponse<T>(
-    val data: T? = null,
-    val error: String? = null,
-    val success: Boolean,
-)
