--- app/src/main/java/com/example/androidcodelabapp/data/repositories/DeveloperRepositoryImpl.kt
@@ -1,104 +1,0 @@
-package com.example.androidcodelabapp.data.repositories
-
-import com.example.androidcodelabapp.data.mappers.toDomain
-import com.example.androidcodelabapp.data.network.responses.NetworkResponse
-import com.example.androidcodelabapp.data.network.api.APIClient
-import com.example.androidcodelabapp.data.network.api.ApiService
-import com.example.androidcodelabapp.data.network.dto.GithubUserDto
-import com.example.androidcodelabapp.data.network.dto.GithubUsersResponseDto
-import com.example.androidcodelabapp.domain.models.GitHubUserResponse
-import com.example.androidcodelabapp.domain.repositories.DeveloperRepository
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-
-class DeveloperRepositoryImpl : DeveloperRepository {
-
-    private val apiClient: APIClient = ApiService.apiClient
-
-    override fun getDevelopers(
-        onResult: (NetworkResponse<GitHubUserResponse>) -> Unit
-    ) {
-        try {
-            apiClient.getAllDevelopers()
-                .enqueue(object : Callback<GithubUsersResponseDto> {
-                    override fun onResponse(
-                        call: Call<GithubUsersResponseDto>,
-                        response: Response<GithubUsersResponseDto>
-                    ) {
-                        if (response.isSuccessful) {
-                            val dtoResponse = response.body()
-                            val domainResponse = dtoResponse?.toDomain()
-                            onResult(
-                                NetworkResponse(
-                                    data = domainResponse,
-                                    success = true
-                                )
-                            )
-
-                        } else {
-                            onResult(
-                                NetworkResponse(
-                                    error = response.errorBody()?.string(),
-                                    success = false
-                                )
-                            )
-                        }
-                    }
-
-                    override fun onFailure(
-                        call: Call<GithubUsersResponseDto>,
-                        t: Throwable
-                    ) {
-                        onResult(
-                            NetworkResponse(
-                                error = t.message,
-                                success = false
-                            )
-                        )
-                    }
-                })
-
-        } catch (e: Exception) {
-            onResult(
-                NetworkResponse(
-                    error = e.message,
-                    success = false
-                )
-            )
-        }
-    }
-
-    override fun getDeveloperProfile(
-        handle: String,
-        onResult: (NetworkResponse<GithubUserDto>) -> Unit
-    ) {
-        try {
-            apiClient.getDeveloperProfile(handle).enqueue(object : Callback<GithubUserDto> {
-                override fun onResponse(
-                    call: Call<GithubUserDto>,
-                    response: Response<GithubUserDto>
-                ) {
-                    if (response.isSuccessful) {
-                        val responseData = response.body()
-                        onResult(NetworkResponse(success = true, data = responseData))
-                    } else {
-                        val error = response.errorBody()?.string()
-                        onResult(NetworkResponse(error = error, success = false))
-                    }
-                }
-
-                override fun onFailure(call: Call<GithubUserDto>, t: Throwable) {
-                    val error = t.message
-                    onResult(NetworkResponse(error = error, success = false))
-                }
-            })
-        } catch (e: Exception) {
-            val error = e.message
-            onResult(NetworkResponse(error = error, success = false))
-
-
-        }
-
-    }
-}
