--- app/src/main/java/com/example/androidcodelabapp/domain/models/GitHubUserResponse.kt
@@ -1,5 +1,0 @@
-package com.example.androidcodelabapp.domain.models
-
-data class GitHubUserResponse(
-    val githubUsers: ArrayList<GitHubUser> = arrayListOf()
-)
