--- app/src/main/java/com/example/androidcodelabapp/presentation/developer_profile/SingleDeveloperContract.kt
@@ -1,8 +1,0 @@
-package com.example.androidcodelabapp.presentation.developer_profile
-
-import com.example.androidcodelabapp.data.network.dto.GithubUserDto
-
-interface SingleDeveloperContract {
-    fun showDeveloperProfile(profile: GithubUserDto)
-    fun showError(s: String)
-}
