--- app/src/main/java/com/example/androidcodelabapp/presentation/developers/DevelopersContract.kt
@@ -1,8 +1,0 @@
-package com.example.androidcodelabapp.presentation.developers;
-
-import com.example.androidcodelabapp.domain.models.GitHubUserResponse
-
-interface DevelopersContract {
-    fun showDevelopers(response: GitHubUserResponse)
-    fun showError(s: String)
-}
