--- app/src/main/java/com/example/androidcodelabapp/util/Constants.kt
@@ -1,5 +1,0 @@
-package com.example.androidcodelabapp.util
-
-object Constants{
-    const val BASEURL:String = "https://api.github.com/"
-}
