+++ app/src/main/java/com/kamira/githubusers/data/network/dto/GithubUsersResponseDto.kt
@@ -1,0 +1,8 @@
+package com.kamira.githubusers.data.network.dto
+
+import com.google.gson.annotations.SerializedName
+
+data class GithubUsersResponseDto(
+    @SerializedName("items")
+    val githubUsers: ArrayList<GithubUserDto> = arrayListOf()
+)
