+++ app/src/main/java/com/kamira/githubusers/data/network/responses/NetworkResponse.kt
@@ -1,0 +1,7 @@
+package com.kamira.githubusers.data.network.responses
+
+data class NetworkResponse<T>(
+    val data: T? = null,
+    val error: String? = null,
+    val success: Boolean,
+)
