+++ app/src/main/java/com/kamira/githubusers/domain/models/GitHubUser.kt
@@ -1,0 +1,12 @@
+package com.kamira.githubusers.domain.models
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class GitHubUser(
+    val profileImage: String?,
+    val userName: String?,
+    val profile: String?,
+    val organization: String?
+) : Parcelable
