+++ app/src/main/java/com/kamira/githubusers/domain/models/GitHubUserResponse.kt
@@ -1,0 +1,5 @@
+package com.kamira.githubusers.domain.models
+
+data class GitHubUserResponse(
+    val githubUsers: ArrayList<GitHubUser> = arrayListOf()
+)
