+++ app/src/main/java/com/kamira/githubusers/presentation/developer_profile/SingleDeveloperContract.kt
@@ -1,0 +1,8 @@
+package com.kamira.githubusers.presentation.developer_profile
+
+import com.kamira.githubusers.data.network.dto.GithubUserDto
+
+interface SingleDeveloperContract {
+    fun showDeveloperProfile(profile: GithubUserDto)
+    fun showError(s: String)
+}
