+++ app/src/main/java/com/kamira/githubusers/presentation/developers/DevelopersContract.kt
@@ -1,0 +1,8 @@
+package com.kamira.githubusers.presentation.developers;
+
+import com.kamira.githubusers.domain.models.GitHubUserResponse
+
+interface DevelopersContract {
+    fun showDevelopers(response: GitHubUserResponse)
+    fun showError(s: String)
+}
