+++ app/src/main/java/com/kamira/githubusers/util/CheckNetworkConnection.java
@@ -1,0 +1,26 @@
+package com.kamira.githubusers.util;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.Network;
+import android.net.NetworkCapabilities;
+
+public class CheckNetworkConnection {
+
+    private final Context context;
+
+    public CheckNetworkConnection(Context mContext) {
+        this.context = mContext;
+    }
+
+    public boolean isConnected() {
+        ConnectivityManager cm =
+                (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        if (cm == null) return false;
+        Network network = cm.getActiveNetwork();
+        if (network == null) return false;
+        NetworkCapabilities capabilities = cm.getNetworkCapabilities(network);
+        if (capabilities == null) return false;
+        return capabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET);
+    }
+}
