+++ app/src/main/java/com/kamira/githubusers/util/Constants.kt
@@ -1,0 +1,5 @@
+package com.kamira.githubusers.util
+
+object Constants{
+    const val BASEURL:String = "https://api.github.com/"
+}
