+++ app/src/main/res/layout/activity_detail.xml
@@ -12,7 +12,18 @@
         android:layout_height="?attr/actionBarSize"
         android:background="@android:color/white"
         app:navigationIconTint="@android:color/black"
-        app:titleTextColor="@android:color/black" />
+        app:titleTextColor="@android:color/black">
+    <TextView
+        android:id="@+id/toolbarTitle"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:text="@string/profile"
+        android:textColor="@android:color/black"
+        android:textSize="18sp"
+        android:textStyle="bold"/>
+
+    </com.google.android.material.appbar.MaterialToolbar>
 
     <!-- CONTENT -->
     <ScrollView
