+++ app/src/main/res/values/strings.xml
@@ -8,4 +8,5 @@
     <string name="username">Username:</string>
     <string name="organization">Organization:</string>
     <string name="andela">Andela</string>
+    <string name="profile">Profile</string>
 </resources>
