+++ settings.gradle
@@ -14,5 +14,5 @@ dependencyResolutionManagement {
     }
 }
 
-rootProject.name = "AndroidLevelupProject"
+rootProject.name = "Github Users"
 include(":app")
